requirements.txt
# FZU_SentimentAnalysis - 细粒度情感分析系统依赖包

# Web 框架
streamlit>=1.28.0

# 深度学习
torch>=2.0.0
transformers>=4.30.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.0.0

# 可视化
plotly>=5.14.0

# 工具
pathlib>=1.0.1
requests>=2.31.0
Pillow>=10.2.0
tomli>=2.0.1

# 文档处理
pdfplumber>=0.10.0
python-docx>=1.0.0
markdown>=3.4.0
beautifulsoup4>=4.12.0
